Avoid system() as this seems to crash net_applet.
In what would appear to be related to threads and gtk3, calling system()
here appears to crash net_applet().

We can avoid it by simply reading the network config.

I'm not sure if a global $net variable is available here
or if things could be re-engineered to pass this object in,
but this seems like a quick and easy fix.

mga#11929