ifcfg: Do not write the NM_CONTROLLED flag unless we know it's value.
If it's not read from the config, it will be read as undef, but written
back as 'no'. This changes the behaviour where no setting means
'automatic' - i.e. if NM is installed, use it, otherwise don't.

If the user wants to be specific, then they make the consious choice
to tick the box in drakx-net.

mga#6675 mga#9261